+++ src/main/java/com/example/met/dto/TimeSpentSummary.java
@@ -16,8 +16,9 @@ public class TimeSpentSummary {
     private String totalAssignedTime;    // Format: "HH:mm"
     private String totalCombinedTime;    // Format: "HH:mm"
 
-    private int totalOnHoldMinutes;
-    private int totalInProgressMinutes;
-    private int totalAssignedMinutes;
-    private int totalCombinedMinutes;
+    private long totalOnHoldMinutes;
+    private long totalInProgressMinutes;
+    private long totalAssignedMinutes;
+    private long totalCombinedMinutes;
+
 }
